Documentation/ spliting class
- added documentation
- splited search file
- added dropdown for status and description field

NOT TESTED due to issues with dockerfile
