+++ web/src/components/views/SearchPage.jsx
@@ -1,6 +1,7 @@
 import React, { Component } from 'react';
 import { Link } from 'react-router-dom';
 import {fetchJson, sendJson} from '../common/Backend'
+import  {Table} from 'SearchTable.jsx';
 
 const defaultPageSize = 4;
 const defaultSearchString = "advanced/?q=*";
@@ -123,14 +124,31 @@ class AdvancedSearch extends Component {
               <span className ="input-group-addon primary">
                 Status
               </span>
-              <input
+              <select
                 className="form-control"
-                type="text"
                 id="status"
                 name="status"
                 onChange={(value) => this.props.ChangeStateStatus(value.target.value)}
-              />
+              >
+                <option>DONE</option>
+                <option>pending</option>
+                <option>in progress</option>
+              </select>
             </div>
+          </div>
+        </div>
+        <div className="col-md-6">
+          <div className="input-group form-inline panel">
+            <span className ="input-group-addon primary">
+              Description
+            </span>
+            <input
+              className="form-control"
+              type="text"
+              id="description"
+              name="description"
+              onChange={(value) => this.props.changeStateDescription(value.target.value)}
+            />
           </div>
         </div>
       </div>
@@ -165,6 +183,7 @@ class Search extends Component{
                 changeStateTo={(to) => this.props.changeStateTo(to)}
                 ChangeStateStatus={(status) => this.props.ChangeStateStatus(status)}
                 changeStateTags={(tags) => this.props.changeStateTags(tags)}
+                changeStateDescription={(desc) => this.props.changeStateDescription(desc)}
               />
             </form>
             <a onClick={() => this.toggle()}  className="clickable-text col-md-2">
@@ -192,300 +211,7 @@ class Search extends Component{
   }
 }
 
-class Table extends Component {
-  constructor(props) {
-    super(props);
-    this.state = {
-      results : [],
-      searchString : defaultSearchString,
-      pageSize : defaultPageSize,
-      pageNumber : 0,
-      numberOfResults : 0,
-      dirty : false,
-      hasPrev: false,
-      hasNext: false,
-      lastString: ""
-    };
 
-    // Get Data from Elasticsearch and put it in this.state.results
-    var that = this;
-    this.getData(this.state.searchString);
-  }
-
-    getData(searchString){
-    var that = this;
-    fetchJson('/api/projects/search/'+
-             this.props.searchString +'&'+
-              'offset='+ this.state.pageNumber+'&'+
-              'count='+ this.state.pageSize)
-    .then(function(data) {
-      if(data != null){
-        that.setState({
-          numberOfResults : data.total,
-          hasNext : data.total > defaultPageSize,
-        });
-        data = data.hits;
-        var validatedData = [];
-        for(var i = 0;i<data.length;i++){
-          if (data[i]._source!=null&&
-              data[i]._source.title!=null &&
-              data[i]._source.authors!=null&&
-              data[i]._source.date_creation!=null&&
-              data[i]._source.description!=null&&
-              data[i]._source.status!=null&&
-              data[i]._id!=null){
-                var export_data = data[i]._source
-                export_data["_id"] = data[i]._id
-                validatedData.push(export_data)
-          }
-        }
-        that.setState({
-           results : validatedData,
-        });
-      }
-    });
-  }
-
-  fitLength(string, maxLength){
-    if(string.length>maxLength){
-      return string.substring(0, maxLength) + " ...";
-    }
-    else return string;
-  }
-
-  authorsArrayToNameString(authors){
-    var names = [];
-    for(var i = 0 ;i < authors.length;i++){
-      names.push(authors[i].name);
-    }
-    return names.join(", ");
-  }
-
-  renderLine(result,index){
-
-    var id = result._id;
-    var title = result.title;
-    var authors = result.authors;
-    var description = result.description;
-    var status = result.status;
-    var date_creation = result.date_creation;
-
-    authors = this.authorsArrayToNameString(authors);
-    var shortenedDescription = this.fitLength(description, 100);
-    return(
-        <tr key ={"result"+index}>
-          <td>
-            <Link to={"/projects/" + id} className="table-project-name">
-              <a className="table-project-name" >
-                <u>{title}</u>
-              </a>
-            </Link>
-          </td>
-          <td> {authors}</td>
-          <td> {status} </td>
-          <td data-toggle="tooltip" title={description}> {shortenedDescription} </td>
-          <td> {date_creation} </td>
-          <td> {this.shuffleBookmarks()} </td>
-        </tr>
-    );
-  }
-
-  shuffleBookmarks() {
-    var x = Math.floor((Math.random() * 2) + 1);
-    if (x == 1) {
-      return <i className="fa fa-bookmark" aria-hidden="true" />;
-    }
-    else {
-      return <i className="fa fa-bookmark-o" aria-hidden="true" />;
-    }
-  }
-
-  renderLines(results){
-    var lines = [];
-    for (var i = 0; i < results.length; i++) {
-      lines.push(this.renderLine(results[i],i));
-    }
-    return(lines);
-  }
-
-  /*
-   * Load a new Page.
-   * Set changeIndex to the number of pages you want to go forward
-   * or backward using positive or negative integers respectively.
-   */
-  newPage(changeIndex){
-    var newPageIndex = this.state.pageNumber + (changeIndex * this.state.pageSize);
-
-    if( newPageIndex>0){
-      this.setState({
-        dirty : true,
-        pageNumber : newPageIndex,
-      });
-    } else {
-      this.setState({
-        dirty : true,
-        pageNumber : 0,
-      })
-    }
-  }
-
-  renderNextButton(){
-    if(this.state.hasNext){
-      return(
-        <span className="input-group-button">
-          <button className="btn btn-secondary" onClick = {() => {this.newPage(1)}}>
-            next
-          </button>
-        </span>
-      );
-    } else {
-      return(
-        <span className="input-group-button">
-          <button className="btn btn-secondary disabled">
-            next
-          </button>
-        </span>
-      );
-    }
-  }
-
-  renderPrevButton(){
-    if(this.state.hasPrev){
-      return(
-        <span className="input-group-button">
-          <button className="btn btn-secondary" onClick = {() => {this.newPage(-1)}}>
-            prev
-          </button>
-        </span>
-      );
-    } else {
-      return(
-        <span className="input-group-button">
-          <button className="btn btn-secondary disabled">
-            prev
-          </button>
-        </span>
-      );
-    }
-  }
-
-  renderTable(results){
-    if(this.state.lastString != this.props.searchString){
-        this.getData(this.props.searchString)
-        this.setState({lastString: this.props.searchString})
-
-    }
-    if(results.length > 0){
-      return(
-        <div>
-          <div className="row">
-            <div className="col-2">
-              <label htmlFor="n-results"> Show Results:&nbsp;</label>
-              <select className="selectpicker"
-                      id="n-results"
-                      onChange={event => this.setState({
-                         pageSize : parseInt(event.target.value, 10),
-                         pageNumber : 0,
-                         dirty : true,
-                       })}
-                      value={this.state.pageSize}>
-                <option value="4">4</option>
-                <option value="10">10</option>
-                <option value="20">20</option>
-              </select>
-            </div>
-          </div>
-          <div className="row">
-            <div className="table-container">
-              <table className="table">
-                <thead className="thead-default">
-                  <tr>
-                    <th className="col-xs-3">Project Name</th>
-                    <th className="col-xs-2">Authors</th>
-                    <th className="col-xs-1">Status</th>
-                    <th className="col-xs-3">Description</th>
-                    <th className="col-xs-2">Date</th>
-                    <th className="col-xs-1">Fav</th>
-                  </tr>
-                </thead>
-                <tbody>
-                  {this.renderLines(results)}
-                </tbody>
-              </table>
-            </div>
-          </div>
-          <div className="row">
-            <div className="col-12">
-              <div className="centered-pagination">
-                <div>
-                  {this.renderPrevButton()}
-                  {this.renderNextButton()}
-                </div>
-              </div>
-            </div>
-          </div>
-        </div>
-      );
-    } else {
-      return null;
-    }
-  }
-
-  componentDidUpdate(){
-    //Load current table page
-    if(this.state.dirty){
-      var that = this;
-      this.setState({
-        dirty : false,
-      });
-
-      fetchJson('/api/projects/search/'+
-                this.props.searchString +'&'+
-                'offset='+ this.state.pageNumber+'&'+
-                'count='+ this.state.pageSize)
-      .then(function(data) {
-        if(data != null){
-          that.setState({
-            numberOfResults : data.total,
-          });
-          data = data.hits;
-          var validatedData = [];
-          for(var i = 0;i<data.length;i++){
-            if (data[i]._source!=null&&
-                data[i]._source.title!=null &&
-                data[i]._source.authors!=null&&
-                data[i]._source.date_creation!=null&&
-                data[i]._source.description!=null&&
-                data[i]._source.status!=null&&
-                data[i]._id!=null){
-                  var export_data = data[i]._source
-                  export_data["_id"] = data[i]._id
-                  validatedData.push(export_data)
-            }
-          }
-          var hasNext = true;
-          if(that.state.numberOfResults-(that.state.pageSize+that.state.pageNumber)<=0){
-            hasNext = false;
-          }
-          var hasPrev = true;
-          if(that.state.pageNumber===0){
-            hasPrev = false;
-          }
-          that.setState({
-             results : validatedData,
-             hasPrev : hasPrev,
-             hasNext : hasNext,
-          });
-        }
-      });
-    }
-  }
-
-  render() {
-
-    return this.renderTable(this.state.results);
-  }
-}
 
 export default class SearchPage extends Component {
   constructor(){
@@ -504,6 +230,10 @@ export default class SearchPage extends Component {
     };
   }
 
+/*
+* The filter function is checking whitch filters are set and is building the searchquery
+*/
+
   filter(){
 
       var searchstring="advanced/?q=";
@@ -514,7 +244,6 @@ export default class SearchPage extends Component {
         filter_set++;
       }
 
-
       if(this.state.filter_date_from !== "1900-01-01" || this.state.filter_date_to!== "2050-06-06"){
         if(filter_set > 0){
           searchstring = searchstring.concat(" AND ");
@@ -524,7 +253,6 @@ export default class SearchPage extends Component {
         filter_set++;
       }
 
-
       if(this.state.filter_author != ""){
         if(filter_set > 0){
           searchstring = searchstring.concat(" AND ");
@@ -532,7 +260,6 @@ export default class SearchPage extends Component {
         }
         searchstring = searchstring.concat("(authors.name: ", this.state.filter_author, "*)");
         filter_set++;
-
       }
       if(this.state.filter_tags != ""){
         if(filter_set > 0){
@@ -557,12 +284,14 @@ export default class SearchPage extends Component {
 
 
   /*
-    functions to get the state of the advanced search via the search element, the value of the input fields is given to the <table> and filtered
-  */
+   * various functions to get the state of the advanced search via the search element, the value of the input fields are
+   * combined to get the searchquery string
+   */
   changeStateAuthor(author){
     this.setState({filter_author: author });
   }
 
+  //Sets Date to 1900-01-01 it date is empty to get filters working if one date field is set
   changeStateFrom(from){
     if(from == ""){
       this.setState({filter_date_from: "1900-01-01" });
@@ -571,6 +300,7 @@ export default class SearchPage extends Component {
     }
   }
 
+  //Sets Date to 2050-06-06 it date is empty to get filters working if one date field is set
   changeStateTo(to){
     if(to == ""){
       this.setState({filter_date_to: "2050-06-06" });
@@ -595,18 +325,31 @@ export default class SearchPage extends Component {
     this.setState({advanced: !this.state.advanced})
   }
 
+  changeStateDescription(desc){
+    this.setState({description: desc})
+  }
+
+
   getSearchString(str){
     this.setState({simple_searchstring: str})
   }
+
+
+/*
+ *This function builds the searchquery for the simplesearch
+ */
 
   simplesearch(){
     var searchString = "simple/?q=";
     searchString = searchString.concat(this.state.simple_searchstring + "*");
     return searchString;
-
   }
 
 
+/*
+ *The renderfunction is checking if advanced search or simple search is needed and is calling the function to create the searchString
+ *which is given to the table class to get the data from the backend
+ */
   render() {
     var searchString
 
@@ -636,6 +379,7 @@ export default class SearchPage extends Component {
                   changeStateTo={(to) => this.changeStateTo(to)}
                   ChangeStateStatus={(status) => this.changeStateStatus(status)}
                   changeStateTags={(tags) => this.changeStateTags(tags)}
+                  changeStateDescription = {(desc) => this.changeStateDescription(desc)}
                   changeStateAdvanced = {(advanced) => this.changeStateAdvanced(advanced)}
                   getSearchString = {(str)=> this.getSearchString(str)}
                   simple_searchString= {this.state.simple_searchString
@@ -644,6 +388,7 @@ export default class SearchPage extends Component {
                 <hr className="horizontal-divider"/>
                 <Table
                 searchString  = {searchString}
+                advanced = {this.state.advanced}
                 />
             </div>
           </div>
